+++ concrete_datastore/api/v1/filters.py
@@ -2,11 +2,11 @@
 import uuid
 import functools
 
-from django.db.models import Q, CharField, TextField, ForeignKey
-
+from django.db.models import Q, CharField, TextField, ForeignKey, BooleanField
+from django.db.models.fields.related import ManyToManyField
+from django.contrib.auth import get_user_model
 from rest_framework.exceptions import ValidationError
 from rest_framework.filters import BaseFilterBackend
-
 from concrete_datastore.api.v1.datetime import format_datetime, ensure_pendulum
 
 # target_field type is date or datetime or int or float
@@ -48,7 +48,35 @@ def convert_type(string, field_type, close_period=True):
         return int(string)
 
 
-class FilterUserByLevel(BaseFilterBackend):
+class CustomShemaOperationParameters:
+    def return_if_not_details(self, view, value, extra_condition=True):
+        if view.detail is False and extra_condition:
+            return value
+        return []
+
+
+class FilterUserByLevel(BaseFilterBackend, CustomShemaOperationParameters):
+    def get_schema_operation_parameters(self, view):
+        params = [
+            {
+                'name': 'level',
+                'required': False,
+                'in': 'query',
+                'schema': {'type': 'string'},
+            },
+            {
+                'name': 'atleast',
+                'required': False,
+                'in': 'query',
+                'schema': {'type': 'string'},
+            },
+        ]
+        return self.return_if_not_details(
+            view=view,
+            value=params,
+            extra_condition=view.model_class == get_user_model(),
+        )
+
     def filter_queryset(self, request, queryset, view):
         # Only appyable on user models
         if view.basename.lower() != 'user':
@@ -67,7 +95,24 @@ class FilterUserByLevel(BaseFilterBackend):
         return queryset
 
 
-class FilterSupportingOrBackend(BaseFilterBackend):
+class FilterSupportingOrBackend(
+    BaseFilterBackend, CustomShemaOperationParameters
+):
+    def get_schema_operation_parameters(self, view):
+        params = [
+            {
+                'name': f'{field_name}__in',
+                'required': False,
+                'in': 'query',
+                'description': 'List of values separated by comma',
+                'schema': {'type': 'string'},
+            }
+            for field_name in getattr(view, 'filterset_fields', ())
+            if type(view.model_class._meta.get_field(field_name))
+            != ManyToManyField
+        ]
+        return self.return_if_not_details(view=view, value=params)
+
     def filter_queryset(self, request, queryset, view):
         query_params = request.query_params
         filterset_fields = getattr(view, 'filterset_fields', ())
@@ -80,6 +125,19 @@ class FilterSupportingOrBackend(BaseFilterBackend):
                 continue
 
             values = query_params.get(param).split(',')
+            if isinstance(
+                queryset.model._meta.get_field(param.replace('__in', '')),
+                BooleanField,
+            ):
+                if set(values).difference(['True', 'False', 'None']):
+                    raise ValidationError(
+                        {
+                            "message": (
+                                f"{param}: {values} must contain olny 'True', "
+                                "'False' and/or 'None' (case sensitive)"
+                            )
+                        }
+                    )
             custom_filter = {param: values}
             if q_object is None:
                 q_object = Q(**custom_filter)
@@ -91,7 +149,25 @@ class FilterSupportingOrBackend(BaseFilterBackend):
         return queryset.filter(q_object)
 
 
-class FilterSupportingContainsBackend(BaseFilterBackend):
+class FilterSupportingContainsBackend(
+    BaseFilterBackend, CustomShemaOperationParameters
+):
+    def get_schema_operation_parameters(self, view):
+        params = [
+            {
+                'name': f'{field_name}__contains',
+                'required': False,
+                'in': 'query',
+                'schema': {'type': 'string'},
+            }
+            for field_name in getattr(view, 'filterset_fields', ())
+            if isinstance(
+                view.model_class._meta.get_field(field_name),
+                (CharField, TextField),
+            )
+        ]
+        return self.return_if_not_details(view=view, value=params)
+
     def filter_queryset(self, request, queryset, view):
         query_params = request.query_params
         filterset_fields = getattr(view, 'filterset_fields', ())
@@ -120,7 +196,26 @@ class FilterSupportingContainsBackend(BaseFilterBackend):
         return queryset.filter(q_object)
 
 
-class FilterSupportingEmptyBackend(BaseFilterBackend):
+class FilterSupportingEmptyBackend(
+    BaseFilterBackend, CustomShemaOperationParameters
+):
+    def get_schema_operation_parameters(self, view):
+        params = [
+            {
+                'name': f'{field_name}__isempty',
+                'required': False,
+                'in': 'query',
+                'description': 'True or False',
+                'schema': {'type': 'boolean'},
+            }
+            for field_name in getattr(view, 'filterset_fields', ())
+            if isinstance(
+                view.model_class._meta.get_field(field_name),
+                (CharField, TextField),
+            )
+        ]
+        return self.return_if_not_details(view=view, value=params)
+
     def filter_queryset(self, request, queryset, view):
         query_params = request.query_params
         filterset_fields = getattr(view, 'filterset_fields', ())
@@ -151,7 +246,25 @@ class FilterSupportingEmptyBackend(BaseFilterBackend):
         return queryset.filter(q_object)
 
 
-class FilterSupportingRangeBackend(BaseFilterBackend):
+class FilterSupportingRangeBackend(
+    BaseFilterBackend, CustomShemaOperationParameters
+):
+    def get_schema_operation_parameters(self, view):
+        params = [
+            {
+                'name': f'{field_name}__range',
+                'required': False,
+                'in': 'query',
+                'description': 'A range of values separated by comma',
+                'schema': {'type': 'string'},
+            }
+            for field_name in getattr(view, 'filterset_fields', ())
+            + ('creation_date', 'modification_date')
+            if view.model_class._meta.get_field(field_name).get_internal_type()
+            in RANGEABLE_TYPES
+        ]
+        return self.return_if_not_details(view=view, value=params)
+
     def filter_queryset(self, request, queryset, view):
         query_params = request.query_params
         filterset_fields = getattr(view, 'filterset_fields', ()) + (
@@ -218,7 +331,37 @@ class FilterSupportingRangeBackend(BaseFilterBackend):
         return queryset.filter(q_object)
 
 
-class FilterSupportingComparaisonBackend(BaseFilterBackend):
+class FilterSupportingComparaisonBackend(
+    BaseFilterBackend, CustomShemaOperationParameters
+):
+    def get_schema_operation_parameters(self, view):
+        suffixes_map = {
+            'gte': 'get the values greater than or equal to a given value',
+            'lte': 'get the values less than or equal to a given value',
+            'gt': 'get the values greater than to a given value',
+            'lt': 'get the values less than to a given value',
+        }
+        params = []
+        for key, description in suffixes_map.items():
+            params.extend(
+                [
+                    {
+                        'name': f'{field_name}__{key}',
+                        'required': False,
+                        'in': 'query',
+                        'description': description,
+                        'schema': {'type': 'string'},
+                    }
+                    for field_name in getattr(view, 'filterset_fields', ())
+                    + ('creation_date', 'modification_date')
+                    if view.model_class._meta.get_field(
+                        field_name
+                    ).get_internal_type()
+                    in RANGEABLE_TYPES
+                ]
+            )
+        return self.return_if_not_details(view=view, value=params)
+
     def filter_queryset(self, request, queryset, view):
         query_params = request.query_params
         filterset_fields = getattr(view, 'filterset_fields', ()) + (
@@ -267,7 +410,23 @@ class FilterSupportingComparaisonBackend(BaseFilterBackend):
         return queryset.filter(q_object)
 
 
-class FilterSupportingForeignKey(BaseFilterBackend):
+class FilterSupportingForeignKey(
+    BaseFilterBackend, CustomShemaOperationParameters
+):
+    def get_schema_operation_parameters(self, view):
+        params = [
+            {
+                'name': f'{field_name}_uid',
+                'required': False,
+                'in': 'query',
+                'description': 'UID of the FK',
+                'schema': {'type': 'string'},
+            }
+            for field_name in getattr(view, 'filterset_fields', ())
+            if type(view.model_class._meta.get_field(field_name)) == ForeignKey
+        ]
+        return self.return_if_not_details(view=view, value=params)
+
     def filter_queryset(self, request, queryset, view):
         query_params = request.query_params
         filterset_fields = getattr(view, 'filterset_fields', ())
@@ -302,7 +461,23 @@ class FilterSupportingForeignKey(BaseFilterBackend):
         return queryset.filter(q_object)
 
 
-class FilterForeignKeyIsNullBackend(BaseFilterBackend):
+class FilterForeignKeyIsNullBackend(
+    BaseFilterBackend, CustomShemaOperationParameters
+):
+    def get_schema_operation_parameters(self, view):
+        params = [
+            {
+                'name': f'{field_name}__isull',
+                'required': False,
+                'in': 'query',
+                'schema': {'type': 'boolean'},
+            }
+            for field_name in getattr(view, 'filterset_fields', ())
+            if type(view.model_class._meta.get_field(field_name))
+            in (ForeignKey, ManyToManyField)
+        ]
+        return self.return_if_not_details(view=view, value=params)
+
     def filter_queryset(self, request, queryset, view):
         query_params = request.query_params
         filterset_fields = getattr(view, 'filterset_fields', ())
@@ -317,7 +492,8 @@ class FilterForeignKeyIsNullBackend(BaseFilterBackend):
             if field_name not in filterset_fields:
                 continue
             param_value = True if query_params.get(param) == 'true' else False
-            if type(queryset.model._meta.get_field(field_name)) == ForeignKey:
+            cleaned_param_type = queryset.model._meta.get_field(field_name)
+            if type(cleaned_param_type) in (ForeignKey, ManyToManyField):
                 custom_filter = {param: param_value}
             else:
                 continue
@@ -330,3 +506,59 @@ class FilterForeignKeyIsNullBackend(BaseFilterBackend):
             return queryset
 
         return queryset.filter(q_object)
+
+
+class FilterSupportingManyToMany(
+    BaseFilterBackend, CustomShemaOperationParameters
+):
+    def get_schema_operation_parameters(self, view):
+        params = [
+            {
+                'name': f'{field_name}__in',
+                'required': False,
+                'description': 'List of uids separated by comma',
+                'in': 'query',
+                'schema': {'type': 'string'},
+            }
+            for field_name in getattr(view, 'filterset_fields', ())
+            if type(view.model_class._meta.get_field(field_name))
+            == ManyToManyField
+        ]
+        return self.return_if_not_details(view=view, value=params)
+
+    def filter_queryset(self, request, queryset, view):
+        query_params = request.query_params
+        filterset_fields = getattr(view, 'filterset_fields', ())
+        q_object = None
+        for param in query_params:
+            if not param.endswith('__in'):
+                continue
+            field_name = param.replace('__in', '')
+            if field_name not in filterset_fields:
+                continue
+            cleaned_param_type = queryset.model._meta.get_field(field_name)
+            if not type(cleaned_param_type) == ManyToManyField:
+                continue
+            values = set(query_params.get(param).split(','))
+
+            #:  "value" must be a valid UUID4, otherwise raise ValidationError
+            for value in values:
+                try:
+                    #:  raises ValueError if not UUID4
+                    uuid.UUID(value, version=4)
+                except ValueError:
+                    raise ValidationError(
+                        {
+                            "message": f'{param}: « {value} » is not a valid UUID'
+                        }
+                    )
+
+            custom_filter = {param: values}
+            if q_object is None:
+                q_object = Q(**custom_filter)
+            else:
+                q_object &= Q(**custom_filter)
+        if q_object is None:
+            return queryset
+
+        return queryset.filter(q_object).distinct()